+++ components/ide/jetbrains/gateway-plugin/src/main/resources-stable/META-INF/extensions.xml
@@ -4,6 +4,10 @@
  See License.AGPL.txt in the project root for license information.
 -->
 <idea-plugin>
+    <extensions defaultExtensionNs="com.intellij">
+        <applicationService serviceInterface="io.gitpod.jetbrains.gateway.common.GitpodConnectionHandleFactory"
+                            serviceImplementation="io.gitpod.jetbrains.gateway.stable.StableGitpodConnectionHandleFactory"/>
+    </extensions>
     <extensions defaultExtensionNs="com.jetbrains">
     </extensions>
 </idea-plugin>
